timeline interaction; open problems
