+++ blitz/layout/viewer.py
@@ -4,12 +4,13 @@ from typing import Any, Optional
 import numpy as np
 import pyqtgraph as pg
 from PyQt5.QtCore import Qt, pyqtSignal
-from PyQt5.QtGui import QDropEvent, QMouseEvent
+from PyQt5.QtGui import QDropEvent
 from pyqtgraph import RectROI, mkPen
 
 from .. import settings
 from ..data.load import DataLoader, ImageData
 from ..tools import format_pixel_value, log, wrap_text
+from .widgets import MeasureROI
 
 
 class ImageViewer(pg.ImageView):
@@ -69,7 +70,6 @@ class ImageViewer(pg.ImageView):
         self.crosshair_state = False
         self.toggle_crosshair()
 
-        self.ui.roiPlot.mousePressEvent = self.on_roi_plot_clicked
         self.roi.sigRegionChanged.connect(
             lambda: self.ui.roiPlot.plotItem.vb.autoRange()  # type: ignore
         )
@@ -274,18 +274,6 @@ class ImageViewer(pg.ImageView):
         )
         return current_image, self.image.shape[0]-1, name
 
-    def on_roi_plot_clicked(self, ev) -> None:
-        if isinstance(ev, QMouseEvent):
-            if ev.button() == Qt.MouseButton.MiddleButton:
-                x = self.ui.roiPlot.plotItem.vb.mapSceneToView(  # type: ignore
-                    ev.pos()
-                ).x()
-                index = int(x)
-                index = max(0, min(index, self.image.shape[0]-1))
-                self.setCurrentIndex(index)
-            else:
-                pg.PlotWidget.mousePressEvent(self.ui.roiPlot, ev)
-
     def load_lut_config(self, lut: dict[str, Any]) -> None:
         self.ui.histogram.restoreState(lut)
 
@@ -361,86 +349,3 @@ class ImageViewer(pg.ImageView):
             h_data = self.image[frame_idx, :, y]
             self.v_plot.plot(v_data, x_values, pen='gray')
             self.h_plot.plot(h_data, pen='gray')
-
-
-class MeasureROI(pg.PolyLineROI):
-
-    def __init__(self, view: pg.ViewBox):
-        super().__init__([[0, 0], [0, 20], [10, 10]], closed=True)
-        self.handleSize = 10
-        self.sigRegionChanged.connect(self.update_labels)
-        self.view = view
-
-        self.n_px: int = 1
-        self.px_in_mm: float = 1
-        self.show_in_mm = False
-
-        self.line_labels = []
-        self.angle_labels = []
-
-        self.view.addItem(self)
-        self.setPen(color=(128, 128, 0, 100), width=3)
-        self._visible = True
-        self.toggle()
-
-    def toggle(self) -> None:
-        self._visible = not self._visible
-        self.setVisible(self._visible)
-        for label in self.line_labels:
-            label.setVisible(self._visible)
-        for label in self.angle_labels:
-            label.setVisible(self._visible)
-        if self._visible:
-            self.update_labels()
-
-    def update_labels(self) -> None:
-        self.update_angles()
-        self.update_lines()
-
-    def update_angles(self) -> None:
-        positions = self.getSceneHandlePositions()
-
-        while len(self.angle_labels) > len(positions):
-            label = self.angle_labels.pop()
-            self.view.removeItem(label)
-
-        for i, (_, pos) in enumerate(positions):
-            _, prev_pos = positions[(i - 1) % len(positions)]
-            _, next_pos = positions[(i + 1) % len(positions)]
-            angle = pg.Point(pos - next_pos).angle(
-                pg.Point(pos - prev_pos)
-            )
-            pos = self.view.mapToView(pos)
-            if i < len(self.angle_labels):
-                self.angle_labels[i].setPos(pos.x(), pos.y())
-                self.angle_labels[i].setText(f"{angle:.2f}°")
-            else:
-                angle_label = pg.TextItem(f"{angle:.2f}°")
-                angle_label.setPos(pos.x(), pos.y())
-                self.view.addItem(angle_label)
-                self.angle_labels.append(angle_label)
-
-    def update_lines(self) -> None:
-        positions = self.getSceneHandlePositions()
-
-        while len(self.line_labels) > len(positions):
-            label = self.line_labels.pop()
-            self.view.removeItem(label)
-
-        for i, (_, pos) in enumerate(positions):
-            _, next_pos = positions[(i + 1) % len(positions)]
-            pos = self.view.mapToView(pos)
-            next_pos = self.view.mapToView(next_pos)
-            length = pg.Point(pos - next_pos).length()
-            mid = (pos + next_pos) / 2
-            if self.show_in_mm:
-                length = length * self.px_in_mm / self.n_px
-            pos = self.view.mapToView(pos)
-            if i < len(self.line_labels):
-                self.line_labels[i].setPos(mid.x(), mid.y())
-                self.line_labels[i].setText(f"{length:.2f}")
-            else:
-                angle_label = pg.TextItem(f"{length:.2f}")
-                angle_label.setPos(mid.x(), mid.y())
-                self.view.addItem(angle_label)
-                self.line_labels.append(angle_label)
